requirements.txt
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.4
dnspython==2.8.0
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
pycparser==3.0
pycrypto==2.6.1
pycryptodome==3.23.0
requests==2.32.5
tqdm==4.67.2
urllib3==2.6.3
